Add Level 2, Level 3, app icon and fix music bug
Add Level 2, Level 3 and their corresponding Win Screens.

Level 2 consists in simulating low light environment, so when the light sensor of the phone detects low light environment the level is complete.

Level 3 consists in imitating the movement of a rotation matrix with the phone, so when the phone is rotated the level is complete.

Add app icon. The new icon is the ETSII dome

Fix music bug that prevented music playing when starting the game.

Author: Javier Martínez Ciria <javier.martinez.ciria@alumnos.upm.es>
Date: Sat Jan 14 2023