Fix end game bug, fix app name and minor changes
Fix bug: after last level (Level 3) the current level was updated to level 4, which does not exist. That caused app to crash when trying to load a game which was in level 4.

Fix app name: app name in icon was wrong. It took the first activity (WelcomeScreen) label as name. Label has been changed to GymkETSII so it shows correct app name in app icon.

Minor changes: 
- Indentation
- Comments
- Button color uniformity 
- Language coherence

Author: Javier Jiménez Fernández <j.jfernandez@alumnos.upm.es>
Date: Sun Jan 15 2023